massive files structure cleaning. Integration tests in __tests__
