+++ src/v2/__tests__/mocks/class.spec.ts
@@ -1,4 +1,4 @@
-import { Mockit } from "./Mockit";
+import { Mockit } from "../../Mockit";
 
 class Hellaw {
   public hello(): string {
