+++ src/v2/__tests__/spy/any/single.spec.ts
@@ -1,5 +1,4 @@
-import { Mockit } from "./Mockit";
-import { z } from "zod";
+import { Mockit } from "../../../Mockit";
 
 function hello(...args: any[]) {
   return args;
