+++ src/v2/__tests__/spy/hasBeenCalledX.spec.ts
@@ -1,4 +1,4 @@
-import { Mockit } from "./Mockit";
+import { Mockit } from "../../Mockit";
 
 function hello() {
   return "world";
