+++ src/v2/classMocks/abstract.ts
@@ -1,0 +1,10 @@
+import { FunctionMock } from "../functionMock";
+
+export class AbstractClassMock<T> {
+  constructor(propertiesToMock: Array<keyof T>) {
+    for (const property of propertiesToMock) {
+      const fMock = FunctionMock(property as string);
+      this[property as string] = fMock;
+    }
+  }
+}
